PL dataflow: buffer input to hls::stream first. BIG PERFORMANCE IMPROVEMENT! 0.34x speed now(compared to 0.11x)

+++ pl/include/krnl_col.h
@@ -1,4 +1,4 @@
 #pragma once
 
 void krnl_proc_col(double *descF, bool *P, double *parF,
-                   int64_t descFn, int64_t parFn);
+                   int descFn, int parFn);
